Move config options to client project
